+++ app/src/main/java/com/techcourse/service/TxUserService.java
@@ -1,0 +1,41 @@
+package com.techcourse.service;
+
+import com.techcourse.config.DataSourceConfig;
+import com.techcourse.dao.UserDao;
+import com.techcourse.dao.UserHistoryDao;
+import com.techcourse.domain.User;
+import com.techcourse.domain.UserHistory;
+import org.springframework.transaction.TransactionTemplate;
+
+public class TxUserService implements UserService {
+
+    private final TransactionTemplate transactionTemplate;
+    private final UserDao userDao;
+    private final UserHistoryDao userHistoryDao;
+
+    public TxUserService(final UserDao userDao, final UserHistoryDao userHistoryDao) {
+        this.transactionTemplate = new TransactionTemplate(DataSourceConfig.getInstance());
+        this.userDao = userDao;
+        this.userHistoryDao = userHistoryDao;
+    }
+
+    @Override
+    public User findById(final long id) {
+        return userDao.findById(id);
+    }
+
+    @Override
+    public void insert(final User user) {
+        userDao.insert(user);
+    }
+
+    @Override
+    public void changePassword(final long id, final String newPassword, final String createBy) {
+        transactionTemplate.doTransaction(() -> {
+            final var user = findById(id);
+            user.changePassword(newPassword);
+            userDao.update(user);
+            userHistoryDao.log(new UserHistory(user, createBy));
+        });
+    }
+}
